requirements.txt
Flask~=3.0.3
flask-cors~=3.1.0
pillow~=10.4.0
rembg~=2.0.57
gunicorn~=20.1.0
